cardno change  from double to long

+++ Code/FraudDetection/fraudanalytics-dto/src/main/java/com/masteklabs/fraudanalytics/common/dto/PredictionResult.java
@@ -59,7 +59,7 @@ public class PredictionResult implements Serializable{
 	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
-	public Double getCardNumber() {
+	public Long getCardNumber() {
 		return cardNumber;
 	}
 	public void setCardNumber(Long cardNumber) {
